Refactored to fit archunits needs.
